Добавляет ДЗ#7
