+++ js/main.js
@@ -1,4 +1,1 @@
-import './functions.js';
-import {createPhotos} from './photos.js';
-
-createPhotos();
+import './thumbnail.js';
