+++ js/utils.js
@@ -47,7 +47,28 @@ const formingStringChar = (str, num, addStr) => {
   return secStr + newStr + str;
 };
 
+const checkStringLength = (str, maxLength) => str.length <= maxLength;
+
+const extractNumbers = (str) => {
+  let numberStr = '';
+
+  for (let i = 0; i <= str.length; i++) {
+    const letter = parseInt(str[i], 10);
+    if (!isNaN(letter)) {
+      numberStr += letter;
+    }
+  }
+
+  if (numberStr.length) {
+    return parseInt(numberStr, 10);
+  }
+
+  return NaN;
+};
+
+checkStringLength('function', 5);
+extractNumbers('local2 56gg');
 isPalindrome ('топот');
 formingStringChar ('g', 3, 'qwerty');
 
-export {getIdCreator, getRandomNumber, getRandomArrayElement};
+export { getIdCreator, getRandomNumber, getRandomArrayElement };
